quadcopter controller code for roll pitch and yaw

+++ AML/src/DCM.cpp
@@ -2,8 +2,8 @@
 #include <cmath>
 
 namespace AML {
-	const Matrix33 DCM::rotationAroundX(double theta){
-		double data[3][3] = { 
+	const Matrix33 DCM::rotationAroundX(float theta){
+		float data[3][3] = { 
 			{1.0,	 	0.0, 		0.0},
 			{0.0, cos(theta), sin(theta)},
 			{0.0, -sin(theta), cos(theta)}
@@ -11,16 +11,16 @@ namespace AML {
 
 		return Matrix33(data);
 	}
-	const Matrix33 DCM::rotationAroundY(double phi){
-		double data[3][3] = { 
+	const Matrix33 DCM::rotationAroundY(float phi){
+		float data[3][3] = { 
 			{cos(phi), 0.0, -sin(phi)},
 			{	0.0, 	1.0,	 0.0},
 			{sin(phi), 	0.0, cos(phi)}
 		};
 		return Matrix33(data);
 	}
-	const Matrix33 DCM::rotationAroundZ(double thai){
-		double data[3][3] = { 
+	const Matrix33 DCM::rotationAroundZ(float thai){
+		float data[3][3] = { 
 			{cos(thai), sin(thai),	0.0},
 			{-sin(thai), cos(thai),	0.0},
 			{	0.0, 	0.0,	 	1.0},
@@ -28,12 +28,12 @@ namespace AML {
 		return Matrix33(data);
 	}
 
-	bool isValidDCM(const Matrix33 dcm, double tolerance){
-		double det = determinant(dcm);
+	bool isValidDCM(const Matrix33 dcm, float tolerance){
+		float det = determinant(dcm);
 		//first check it's not a negative value
 		if(det <= 0.0) return false;
 		//then check it's equal to 1
-		double error = det - 1.0;
+		float error = det - 1.0;
 		//absolute value of the error
 		error = error < 0 ? -error : error;
 		//if greater than the limit then not a valid dcm
@@ -49,7 +49,7 @@ namespace AML {
 
 		//calculate the error in the directions of the vectors (supposed to be 0)
 		//i.e. all rows and columns should be independent
-		double error = dot(x_component, y_component);
+		float error = dot(x_component, y_component);
 
 		//reverse the vector to the other direction of the error
 		Vector3 orthogonal_x = x_component - 0.5 * error * y_component;
@@ -64,16 +64,16 @@ namespace AML {
 		dcm = Matrix33(x_norm, y_norm, z_norm);
 	}
 
-	Matrix33 integrateDCM(const Matrix33 &dcm, const Matrix33 &dcmRates, double dt){
+	Matrix33 integrateDCM(const Matrix33 &dcm, const Matrix33 &dcmRates, float dt){
 		Matrix33 dcm_update = dcm + dcmRates*dt;
 		normalizeDCM(dcm_update);
 		return dcm_update;
 	}
 	Matrix33 dcmKinematicsBodyRates(const Matrix33 &dcm, const Vector3 &w_b){
-		const double p = w_b.x;
-		const double q = w_b.y;
-		const double r = w_b.z;
-		const double skew_matrix_data[3][3] = {
+		const float p = w_b.x;
+		const float q = w_b.y;
+		const float r = w_b.z;
+		const float skew_matrix_data[3][3] = {
 			{0.0, -r, q},
 			{r, 0.0, -p},
 			{-q, p, 0.0}
@@ -84,10 +84,10 @@ namespace AML {
 		return result;
 	}
 	Matrix33 dcmKinematicsWorldRates(const Matrix33 &dcm, const Vector3 &w_w){
-		const double p = w_w.x;
-		const double q = w_w.y;
-		const double r = w_w.z;
-		const double skew_matrix_data[3][3] = {
+		const float p = w_w.x;
+		const float q = w_w.y;
+		const float r = w_w.z;
+		const float skew_matrix_data[3][3] = {
 			{0.0, -r, q},
 			{r, 0.0, -p},
 			{-q, p, 0.0}
